+++ src/main/java/seedu/address/storage/JsonAdaptedStudent.java
@@ -11,13 +11,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.lab.LabList;
+import seedu.address.model.lab.exceptions.DuplicateLabException;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.GithubUsername;
 import seedu.address.model.student.Name;
 import seedu.address.model.student.Student;
 import seedu.address.model.student.StudentId;
 import seedu.address.model.student.Telegram;
-import seedu.address.model.student.exceptions.DuplicateLabException;
 import seedu.address.model.tag.Tag;
 
 /**
